requirements.txt
torch
torchvision
Pillow
gradio
gunicorn
